requirements/production.txt
#  Production dependencies
-r base.txt
gunicorn==19.7.1
psycopg2==2.6.2
django-debug-toolbar==1.4
